Move add_round files to rounds folder.
